+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # 更新日志
 
++ V3.0.0b1 2020/10/17
+    + 修复article.get_content()报错
+    + 修正一些注释错误
+    + 增加 video.VideoOnlineMonitor，可监控视频实时在线观看人数和实时更新弹幕，使用websocket
+    + 优化 live.LiveDanmaku
+    + 修改 live.connect_all_livedanmaku() 为 connect_all_LiveDanmaku()
+    + 修改 live.LiveDanmaku.connect(return_task) 参数改为 return_coroutine，为True时返回一个 Coroutine 对象而不是 Task 对象
+
 + V3.0.0b0 2020/10/16
     + live.LiveDanmaku现在可以自动重连了，详见文档
     + live.LiveDanmaku.has_connected()更改为get_connect_status，表示连接状态，0未连接，1已连接，3已正常断开，-1异常断开